+++ src/backend/core/api/client/__init__.py
@@ -1,0 +1,1 @@
+"""People core client API endpoints"""
