+++ src/backend/core/tests/team_accesses/test_api_team_accesses_update.py
@@ -9,7 +9,7 @@ import pytest
 from rest_framework.test import APIClient
 
 from core import factories, models
-from core.api import serializers
+from core.api.client import serializers
 
 pytestmark = pytest.mark.django_db
