+++ src/backend/core/tests/test_api_team_invitations.py
@@ -9,7 +9,7 @@ from rest_framework import status
 from rest_framework.test import APIClient
 
 from core import factories
-from core.api import serializers
+from core.api.client import serializers
 
 pytestmark = pytest.mark.django_db
