+++ src/backend/mailbox_manager/api/client/__init__.py
@@ -1,0 +1,1 @@
+"""Mailbox manager client API module."""
