+++ src/backend/mailbox_manager/tests/api/mailboxes/test_api_mailboxes_create.py
@@ -8,7 +8,6 @@ from logging import Logger
 from unittest import mock
 
 from django.test.utils import override_settings
-from django.utils.translation import gettext_lazy as _
 
 import pytest
 import responses
@@ -19,7 +18,7 @@ from rest_framework.test import APIClient
 from core import factories as core_factories
 
 from mailbox_manager import enums, factories, models
-from mailbox_manager.api import serializers
+from mailbox_manager.api.client import serializers
 
 pytestmark = pytest.mark.django_db
