+++ src/backend/people/resource_server_urls.py
@@ -4,8 +4,8 @@ from django.urls import include, path
 
 from rest_framework.routers import DefaultRouter
 
+from core.api.resource_server import viewsets
 from core.resource_server.urls import urlpatterns as resource_server_urls
-from core.resource_server_api import viewsets
 
 # - Main endpoints
 # Contacts will be added later
